Big upgrade to version 1.0
Additions:
—> Boolean Constraint Propagation (BCP): The propagation of a literal
can create unit clauses. Because each clause needs to be satisfied, we
know that this literal must be true, hence it can also be propagated.

This create a need to differentiate between propagations originated by
the arbitrary assignment of a variable and the ones enforced by the
creation of unit clauses. We shall call the former a decision and the
second just propagation.

Now backtracking involves taking back one decision and all its
propagations.

—> Watch lists.
—> Blocking literals.